formatting

+++ packages/ethereum-contracts/contracts/interfaces/superfluid/ISuperfluidToken.sol
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: AGPLv3
-pragma solidity >= 0.8.4;
+pragma solidity >=0.8.4;
 
 import { ISuperAgreement } from "./ISuperAgreement.sol";
 import { SuperfluidErrors } from "./Definitions.sol";
@@ -16,12 +16,12 @@ interface ISuperfluidToken {
     /**
      * @dev Get superfluid host contract address
      */
-    function getHost() external view returns(address host);
+    function getHost() external view returns (address host);
 
     /**
      * @dev Encoded liquidation type data mainly used for handling stack to deep errors
      *
-     * @custom:note 
+     * @custom:note
      * - version: 1
      * - liquidationType key:
      *    - 0 = reward account receives reward (PIC period)
@@ -38,22 +38,21 @@ interface ISuperfluidToken {
      *************************************************************************/
 
     /**
-    * @dev Calculate the real balance of a user, taking in consideration all agreements of the account
-    * @param account for the query
-    * @param timestamp Time of balance
-    * @return availableBalance Real-time balance
-    * @return deposit Account deposit
-    * @return owedDeposit Account owed Deposit
-    */
-    function realtimeBalanceOf(
-       address account,
-       uint256 timestamp
-    )
-        external view
+     * @dev Calculate the real balance of a user, taking in consideration all agreements of the account
+     * @param account for the query
+     * @param timestamp Time of balance
+     * @return availableBalance Real-time balance
+     * @return deposit Account deposit
+     * @return owedDeposit Account owed Deposit
+     */
+    function realtimeBalanceOf(address account, uint256 timestamp)
+        external
+        view
         returns (
             int256 availableBalance,
             uint256 deposit,
-            uint256 owedDeposit);
+            uint256 owedDeposit
+        );
 
     /**
      * @notice Calculate the realtime balance given the current host.getNow() value
@@ -63,41 +62,38 @@ interface ISuperfluidToken {
      * @return deposit Account deposit
      * @return owedDeposit Account owed Deposit
      */
-    function realtimeBalanceOfNow(
-       address account
-    )
-        external view
+    function realtimeBalanceOfNow(address account)
+        external
+        view
         returns (
             int256 availableBalance,
             uint256 deposit,
             uint256 owedDeposit,
-            uint256 timestamp);
+            uint256 timestamp
+        );
 
     /**
-    * @notice Check if account is critical
-    * @dev A critical account is when availableBalance < 0
-    * @param account The account to check
-    * @param timestamp The time we'd like to check if the account is critical (should use future)
-    * @return isCritical Whether the account is critical
-    */
-    function isAccountCritical(
-        address account,
-        uint256 timestamp
-    )
-        external view
-        returns(bool isCritical);
+     * @notice Check if account is critical
+     * @dev A critical account is when availableBalance < 0
+     * @param account The account to check
+     * @param timestamp The time we'd like to check if the account is critical (should use future)
+     * @return isCritical Whether the account is critical
+     */
+    function isAccountCritical(address account, uint256 timestamp)
+        external
+        view
+        returns (bool isCritical);
 
     /**
-    * @notice Check if account is critical now (current host.getNow())
-    * @dev A critical account is when availableBalance < 0
-    * @param account The account to check
-    * @return isCritical Whether the account is critical
-    */
-    function isAccountCriticalNow(
-        address account
-    )
-        external view
-        returns(bool isCritical);
+     * @notice Check if account is critical now (current host.getNow())
+     * @dev A critical account is when availableBalance < 0
+     * @param account The account to check
+     * @return isCritical Whether the account is critical
+     */
+    function isAccountCriticalNow(address account)
+        external
+        view
+        returns (bool isCritical);
 
     /**
      * @notice Check if account is solvent
@@ -106,12 +102,10 @@ interface ISuperfluidToken {
      * @param timestamp The time we'd like to check if the account is solvent (should use future)
      * @return isSolvent True if the account is solvent, false otherwise
      */
-    function isAccountSolvent(
-        address account,
-        uint256 timestamp
-    )
-        external view
-        returns(bool isSolvent);
+    function isAccountSolvent(address account, uint256 timestamp)
+        external
+        view
+        returns (bool isSolvent);
 
     /**
      * @notice Check if account is solvent now
@@ -119,37 +113,33 @@ interface ISuperfluidToken {
      * @param account The account to check
      * @return isSolvent True if the account is solvent, false otherwise
      */
-    function isAccountSolventNow(
-        address account
-    )
-        external view
-        returns(bool isSolvent);
+    function isAccountSolventNow(address account)
+        external
+        view
+        returns (bool isSolvent);
 
     /**
-    * @notice Get a list of agreements that is active for the account
-    * @dev An active agreement is one that has state for the account
-    * @param account Account to query
-    * @return activeAgreements List of accounts that have non-zero states for the account
-    */
+     * @notice Get a list of agreements that is active for the account
+     * @dev An active agreement is one that has state for the account
+     * @param account Account to query
+     * @return activeAgreements List of accounts that have non-zero states for the account
+     */
     function getAccountActiveAgreements(address account)
-       external view
-       returns(ISuperAgreement[] memory activeAgreements);
+        external
+        view
+        returns (ISuperAgreement[] memory activeAgreements);
 
-
-   /**************************************************************************
-    * Super Agreement hosting functions
-    *************************************************************************/
+    /**************************************************************************
+     * Super Agreement hosting functions
+     *************************************************************************/
 
     /**
      * @dev Create a new agreement
      * @param id Agreement ID
      * @param data Agreement data
      */
-    function createAgreement(
-        bytes32 id,
-        bytes32[] calldata data
-    )
-        external;
+    function createAgreement(bytes32 id, bytes32[] calldata data) external;
+
     /**
      * @dev Agreement created event
      * @param agreementClass Contract address of the agreement
@@ -171,21 +161,16 @@ interface ISuperfluidToken {
     function getAgreementData(
         address agreementClass,
         bytes32 id,
-        uint dataLength
-    )
-        external view
-        returns(bytes32[] memory data);
+        uint256 dataLength
+    ) external view returns (bytes32[] memory data);
 
     /**
      * @dev Create a new agreement
      * @param id Agreement ID
      * @param data Agreement data
      */
-    function updateAgreementData(
-        bytes32 id,
-        bytes32[] calldata data
-    )
-        external;
+    function updateAgreementData(bytes32 id, bytes32[] calldata data) external;
+
     /**
      * @dev Agreement updated event
      * @param agreementClass Contract address of the agreement
@@ -202,34 +187,28 @@ interface ISuperfluidToken {
      * @dev Close the agreement
      * @param id Agreement ID
      */
-    function terminateAgreement(
-        bytes32 id,
-        uint dataLength
-    )
-        external;
+    function terminateAgreement(bytes32 id, uint256 dataLength) external;
+
     /**
      * @dev Agreement terminated event
      * @param agreementClass Contract address of the agreement
      * @param id Agreement ID
      */
-    event AgreementTerminated(
-        address indexed agreementClass,
-        bytes32 id
-    );
+    event AgreementTerminated(address indexed agreementClass, bytes32 id);
 
     /**
      * @dev Update agreement state slot
      * @param account Account to be updated
      *
-     * @custom:note 
+     * @custom:note
      * - To clear the storage out, provide zero-ed array of intended length
      */
     function updateAgreementStateSlot(
         address account,
         uint256 slotId,
         bytes32[] calldata slotData
-    )
-        external;
+    ) external;
+
     /**
      * @dev Agreement account state updated event
      * @param agreementClass Contract address of the agreement
@@ -253,10 +232,8 @@ interface ISuperfluidToken {
         address agreementClass,
         address account,
         uint256 slotId,
-        uint dataLength
-    )
-        external view
-        returns (bytes32[] memory slotData);
+        uint256 dataLength
+    ) external view returns (bytes32[] memory slotData);
 
     /**
      * @notice Settle balance from an account by the agreement
@@ -264,14 +241,10 @@ interface ISuperfluidToken {
      * @param account Account to query.
      * @param delta Amount of balance delta to be settled
      *
-     * @custom:modifiers 
+     * @custom:modifiers
      *  - onlyAgreement
      */
-    function settleBalance(
-        address account,
-        int256 delta
-    )
-        external;
+    function settleBalance(address account, int256 delta) external;
 
     /**
      * @dev Make liquidation payouts (v2)
@@ -283,7 +256,7 @@ interface ISuperfluidToken {
      * @param rewardAmount The amount the rewarded account will receive
      * @param targetAccountBalanceDelta The delta amount the target account balance should change by
      *
-     * @custom:note 
+     * @custom:note
      * - If a bailout is required (bailoutAmount > 0)
      *   - the actual reward (single deposit) goes to the executor,
      *   - while the reward account becomes the bailout account
@@ -293,11 +266,10 @@ interface ISuperfluidToken {
      *   - the targetAccount will pay the rewardAmount
      *   - the liquidator (reward account in PIC period) will receive the rewardAmount
      *
-     * @custom:modifiers 
+     * @custom:modifiers
      *  - onlyAgreement
      */
-    function makeLiquidationPayoutsV2
-    (
+    function makeLiquidationPayoutsV2(
         bytes32 id,
         bytes memory liquidationTypeData,
         address liquidatorAccount,
@@ -306,6 +278,7 @@ interface ISuperfluidToken {
         uint256 rewardAmount,
         int256 targetAccountBalanceDelta
     ) external;
+
     /**
      * @dev Agreement liquidation event v2 (including agent account)
      * @param agreementClass Contract address of the agreement
@@ -317,7 +290,7 @@ interface ISuperfluidToken {
      * @param targetAccountBalanceDelta The amount the sender account balance should change by
      * @param liquidationTypeData The encoded liquidation type data including the version (how to decode)
      *
-     * @custom:note 
+     * @custom:note
      * Reward account rule:
      * - if the agreement is liquidated during the PIC period
      *   - the rewardAmountReceiver will get the rewardAmount (remaining deposit), regardless of the liquidatorAccount
@@ -350,8 +323,8 @@ interface ISuperfluidToken {
      * NOTE: solidity-coverage not supporting it
      *************************************************************************/
 
-     /// @dev The msg.sender must be host contract
-     //modifier onlyHost() virtual;
+    /// @dev The msg.sender must be host contract
+    //modifier onlyHost() virtual;
 
     /// @dev The msg.sender must be a listed agreement.
     //modifier onlyAgreement() virtual;
@@ -385,10 +358,7 @@ interface ISuperfluidToken {
      *
      * @custom:deprecated Use AgreementLiquidatedV2 instead
      */
-    event Bailout(
-        address indexed bailoutAccount,
-        uint256 bailoutAmount
-    );
+    event Bailout(address indexed bailoutAccount, uint256 bailoutAmount);
 
     /**
      * @dev Agreement liquidation event (DEPRECATED BY AgreementLiquidatedV2)
@@ -402,7 +372,7 @@ interface ISuperfluidToken {
      *
      * @custom:deprecated Use AgreementLiquidatedV2 instead
      *
-     * @custom:note 
+     * @custom:note
      * Reward account rule:
      * - if bailout is equal to 0, then
      *   - the bondAccount will get the rewardAmount,
