+++ packages/ethereum-contracts/contracts/interfaces/tokens/ISETH.sol
@@ -1,8 +1,7 @@
 // SPDX-License-Identifier: AGPLv3
-pragma solidity >= 0.8.4;
+pragma solidity >=0.8.4;
 
 import { ISuperToken } from "../superfluid/ISuperToken.sol";
-
 
 /**
  * @title Super ETH (SETH) custom token interface
@@ -11,8 +10,10 @@ import { ISuperToken } from "../superfluid/ISuperToken.sol";
 interface ISETHCustom {
     // using native token
     function upgradeByETH() external payable;
+
     function upgradeByETHTo(address to) external payable;
-    function downgradeToETH(uint wad) external;
+
+    function downgradeToETH(uint256 wad) external;
 }
 
 /**
@@ -20,4 +21,6 @@ interface ISETHCustom {
  * @author Superfluid
  */
 // solhint-disable-next-line no-empty-blocks
-interface ISETH is ISETHCustom, ISuperToken {}
+interface ISETH is ISETHCustom, ISuperToken {
+
+}
