+++ packages/ethereum-contracts/contracts/interfaces/utils/IMultiSigWallet.sol
@@ -1,12 +1,15 @@
 // SPDX-License-Identifier: UNLICENSED
 
-pragma solidity >= 0.8.4;
+pragma solidity >=0.8.4;
+
 /**
  * @title Multisig wallet interface
  * @author Superfluid
  */
 interface IMultiSigWallet {
-    function submitTransaction(address destination, uint value, bytes calldata data)
-        external
-        returns (uint transactionId);
+    function submitTransaction(
+        address destination,
+        uint256 value,
+        bytes calldata data
+    ) external returns (uint256 transactionId);
 }
