+++ packages/ethereum-contracts/contracts/libs/CallUtils.sol
@@ -6,7 +6,6 @@ pragma solidity 0.8.14;
  * @author Superfluid
  */
 library CallUtils {
-
     /// @dev Bubble up the revert from the returnedData (supports Panic, Error & Custom Errors)
     /// @notice This is needed in order to provide some human-readable revert message from a call
     /// @param returnedData Response of the call
@@ -19,12 +18,14 @@ library CallUtils {
             assembly {
                 errorSelector := mload(add(returnedData, 0x20))
             }
-            if (errorSelector == bytes4(0x4e487b71) /* `seth sig "Panic(uint256)"` */) {
+            if (
+                errorSelector == bytes4(0x4e487b71) /* `seth sig "Panic(uint256)"` */
+            ) {
                 // case 2: Panic(uint256) (Defined since 0.8.0)
                 // solhint-disable-next-line max-line-length
                 // ref: https://docs.soliditylang.org/en/v0.8.0/control-structures.html#panic-via-assert-and-error-via-require)
                 string memory reason = "CallUtils: target panicked: 0x__";
-                uint errorCode;
+                uint256 errorCode;
                 assembly {
                     errorCode := mload(add(returnedData, 0x24))
                     let reasonWord := mload(add(reason, 0x20))
@@ -34,15 +35,19 @@ library CallUtils {
                     let e1 := add(and(errorCode, 0xf), 0x30)
                     let e2 := shl(8, add(shr(4, and(errorCode, 0xf0)), 0x30))
                     reasonWord := or(
-                        and(reasonWord, 0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff0000),
-                        or(e2, e1))
+                        and(
+                            reasonWord,
+                            0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff0000
+                        ),
+                        or(e2, e1)
+                    )
                     mstore(add(reason, 0x20), reasonWord)
                 }
                 revert(reason);
             } else {
                 // case 3: Error(string) (Defined at least since 0.7.0)
                 // case 4: Custom errors (Defined since 0.8.0)
-                uint len = returnedData.length;
+                uint256 len = returnedData.length;
                 assembly {
                     revert(add(returnedData, 32), len)
                 }
@@ -51,12 +56,16 @@ library CallUtils {
     }
 
     /**
-    * @dev Helper method to parse data and extract the method signature (selector).
-    *
-    * Copied from: https://github.com/argentlabs/argent-contracts/
-    * blob/master/contracts/modules/common/Utils.sol#L54-L60
-    */
-    function parseSelector(bytes memory callData) internal pure returns (bytes4 selector) {
+     * @dev Helper method to parse data and extract the method signature (selector).
+     *
+     * Copied from: https://github.com/argentlabs/argent-contracts/
+     * blob/master/contracts/modules/common/Utils.sol#L54-L60
+     */
+    function parseSelector(bytes memory callData)
+        internal
+        pure
+        returns (bytes4 selector)
+    {
         require(callData.length >= 4, "CallUtils: invalid callData");
         // solhint-disable-next-line no-inline-assembly
         assembly {
@@ -67,8 +76,18 @@ library CallUtils {
     /**
      * @dev Pad length to 32 bytes word boundary
      */
-    function padLength32(uint256 len) internal pure returns (uint256 paddedLen) {
-        return ((len / 32) +  (((len & 31) > 0) /* rounding? */ ? 1 : 0)) * 32;
+    function padLength32(uint256 len)
+        internal
+        pure
+        returns (uint256 paddedLen)
+    {
+        return
+            ((len / 32) +
+                (
+                    ((len & 31) > 0) /* rounding? */
+                        ? 1
+                        : 0
+                )) * 32;
     }
 
     /**
@@ -78,16 +97,23 @@ library CallUtils {
      * | word 1      | word 2           | word 3           | the rest...
      * | data length | bytesData offset | bytesData length | bytesData + padLength32 zeros |
      */
-    function isValidAbiEncodedBytes(bytes memory data) internal pure returns (bool) {
+    function isValidAbiEncodedBytes(bytes memory data)
+        internal
+        pure
+        returns (bool)
+    {
         if (data.length < 64) return false;
-        uint bytesOffset;
-        uint bytesLen;
+        uint256 bytesOffset;
+        uint256 bytesLen;
         // bytes offset is always expected to be 32
-        assembly { bytesOffset := mload(add(data, 32)) }
+        assembly {
+            bytesOffset := mload(add(data, 32))
+        }
         if (bytesOffset != 32) return false;
-        assembly { bytesLen := mload(add(data, 64)) }
+        assembly {
+            bytesLen := mload(add(data, 64))
+        }
         // the data length should be bytesData.length + 64 + padded bytes length
         return data.length == 64 + padLength32(bytesLen);
     }
-
 }
